freda_zhaobiao/requirements.txt
flask>=2.3.0
flask-sqlalchemy>=3.0.0
flask-login>=0.6.0
flask-wtf>=1.1.0
flask-caching>=2.0.0
werkzeug>=2.3.0
openpyxl>=3.1.0
pandas>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
apscheduler>=3.10.0
sqlalchemy>=2.0.0
wtforms>=3.0.0
email-validator>=2.0.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
psycopg2-binary>=2.9.0
